Bernard_HamRadio {xUCINET}	R Documentation

Amateur HAM radio calls by Bernard and colleagues.

Description

Bernard & Killworth, later with the help of Sailer, collected five sets of data on human interactions in bounded groups and on the actors' ability to recall those interactions. In each study they obtained measures of social interaction among all actors, and ranking data based on the subjects' memory of those interactions. This dataset is based on amateur HAM radio calls made over a one-month period. See $ProjectInfo for more information.

Usage


Bernard_HamRadio

Format

An object of class list of length 5.

Source

UCINET

References

See $ProjectInfo$References

Examples


Bernard_HamRadio


[Package xUCINET version 0.0.1.0022 Index]