Borgatti_Camp92 {xUCINET}	R Documentation

Network among participants of a workshop

Description

These data were collected by Steve Borgatti, Russ Bernard, Bert Pelto and Gery Ryan at the 1992 NSF Summer Institute on Research Methods in Cultural Anthropology. This was a 3-week course given to 14 carefully selected participants. Network data were collected at the end of each week. These data were collected at the end of the second week. The data were collected by placing each person's name on a card and asking each respondent to sort the cards in order of how much interaction they had with that person since the beginning of the course (known informally as 'camp'). This results in rank order data in which a '1' indicates the most interaction while a '17' indicates the least interaction. See $ProjectInfo for more information.

Usage


Borgatti_Camp92

Format

An object of class list of length 5.

Source

UCINET

References

See $ProjectInfo$References

Examples


Borgatti_Camp92


[Package xUCINET version 0.0.1.0022 Index]