Borgatti_Cities {xUCINET}	R Documentation

Matrix capturing distances between US cities

Description

This network contain the distance in miles between some of the major cities in the US. The distances are based on how many miles it takes to drive from one city to the other. See $ProjectInfo for more information.

Usage


Borgatti_Cities

Format

An object of class list of length 4.

Source

UCINET

References

See $ProjectInfo$References

Examples


Borgatti_Cities


[Package xUCINET version 0.0.1.0022 Index]