Borgatti_Scientists960 {xUCINET}	R Documentation

Collaboration among 960 scientists.

Description

These data indicate the amount of time (in days, and rounded to the nearest day) that scientists at a large applied research institute worked on projects together. The data are derived from a 2-mode matrix (960 persons by 3154 projects) in which the cells indicated how much time each scientist worked on each project. The 1-mode projection was achieved by the sum of minimums method. For example, if person A spent 1 day on project 1 and 3 days on project 2, and person B spent 5 days on project 1 and 2 days on project 2, the amount of time worked together is 1+2 = 3 days. See $ProjectInfo for more information.

Usage


Borgatti_Scientists960

Format

An object of class list of length 4.

Source

UCINET

References

See $ProjectInfo$References

Examples


Borgatti_Scientists960


[Package xUCINET version 0.0.1.0022 Index]