Burkhardt_GovernmentAgency {xUCINET}	R Documentation

Advice among employees in a government agency collected by Burkhardt and Brass.

Description

Network data capturing who communicated with whom about work matters at a government agency at five points in tie. The data were collected by Burkhardt and Brass on 75 employees. See $ProjectInfo for more information.

Usage


Burkhardt_GovernmentAgency

Format

An object of class list of length 8.

Source

UCINET

References

See $ProjectInfo$References

Examples


Burkhardt_GovernmentAgency


[Package xUCINET version 0.0.1.0022 Index]