Davis_SouthernWomen {xUCINET}	R Documentation

Two-mode network of Southern women attending events.

Description

These data were collected by Davis, Gardner and Gardner in the 1930s. They represent attendance at 14 social events by 18 Southern women. The result is a person-by-event matrix, with people in the rows and events in the columns. A cell (i,j) is 1 if person i attended social event j, and 0 otherwise. (See also Chapter 13 of Borgatti et al., 2022). See $ProjectInfo for more information.

Usage


Davis_SouthernWomen

Format

An object of class list of length 4.

Source

UCINET

References

See $ProjectInfo$References

Examples


Davis_SouthernWomen


[Package xUCINET version 0.0.1.0022 Index]