Freeman_EIES {xUCINET}	R Documentation

Freeman's EIES network dataset

Description

This data arose from an early experiment on computer mediated communication. Fifty academics interested in interdisciplinary research were allowed to contact each other via an Electronic Information Exchange System (EIES). The data collected consisted of all messages sent as well as acquaintance relationships at two time periods (collected via a questionnaire). The data includes the 32 actors who completed the study. In addition, attribute data on primary discipline and number of citations was recorded. See $ProjectInfo for more information.

Usage


Freeman_EIES

Format

An object of class list of length 6.

Source

UCINET

References

See $ProjectInfo$References

[Package xUCINET version 0.0.1.0022 Index]