Hawthorne_BankWiring {xUCINET}	R Documentation

Interactions among employees in a bank wiring room (Hawthorne plant).

Description

These are the observational data on 14 Western Electric (Hawthorne Plant) employees from the bank wiring room first presented in Roethlisberger & Dickson (1939). The data are better known through a scrutiny made of the interactions in Homans (1950), as well as the CONCOR analysis presented in Breiger et al (1975). The employees worked in a single room and include two inspectors (I1 and I3), three solderers (S1, S2 and S3), and nine wiremen or assemblers (W1 to W9). Contains six types of interactions among employees. See $ProjectInfo for more information.

Usage


Hawthorne_BankWiring

Format

An object of class list of length 9.

Source

UCINET

References

See $ProjectInfo$References

Examples


Hawthorne_BankWiring


[Package xUCINET version 0.0.1.0022 Index]