Kapferer_TailorShop {xUCINET}	R Documentation

Interactions among workers in a tailor shop school kids in a school collected by Kapferer.

Description

Bruce Kapferer (1972) observed interactions in a tailor shop in Zambia (then Northern Rhodesia) over a period of ten months. His focus was the changing patterns of alliance among workers during extended negotiations for higher wages. The matrices represent two different types of interaction, recorded at two different times (seven months apart) over a period of one month: 'instrumental' (work- and assistance-related) interactions at the two times; and 'sociational' (friendship, socioemotional) interactions. The data are particularly interesting since an abortive strike occurred after the first set of observations, and a successful strike took place after the second. See $ProjectInfo for more information.

Usage


Kapferer_TailorShop

Format

An object of class list of length 7.

Source

UCINET

References

See $ProjectInfo$References

Examples


Kapferer_TailorShop


[Package xUCINET version 0.0.1.0022 Index]