Knecht_Class12b {xUCINET}	R Documentation

Network among school kids in a school collected by Knecht.

Description

This data was collected in a Dutch school by Andrea Knecht as part of her dissertation (Knecht, 2008). It measures friendship at 4 time points among 26 11-13 year old kids, as well as delinquency and alcohol use (the last was not measured at time 1). It also contains information about whether the kids knew each other from primary education. See $ProjectInfo for more information.

Usage


Knecht_Class12b

Format

An object of class list of length 8.

Source

UCINET

References

See $ProjectInfo$References

Examples


Knecht_Class12b


[Package xUCINET version 0.0.1.0022 Index]