Krackhardt_HighTech {xUCINET}	R Documentation

Network ties among managers in a high-tech company collected by Krackhardt.

Description

These are data collected by David Krackhardt among the managers of a high-tech company. The company manufactured high-tech equipment on the West Coast of the United States and had just over 100 employees with 21 managers. Contains three types of network relations and four attributes. See $ProjectInfo for more information.

Usage


Krackhardt_HighTech

Format

An object of class list of length 6.

Source

UCINET

References

See $ProjectInfo$References

Examples


Krackhardt_HighTech


[Package xUCINET version 0.0.1.0022 Index]