Newcomb_Fraternity {xUCINET}	R Documentation

Network ties collected by Newcomb among a fraternity.

Description

These 15 matrices record weekly sociometric preference rankings from 17 men attending the University of Michigan in the fall of 1956; data from week 9 are missing. A '1' indicates first preference, and no ties were allowed. The men were recruited to live in off-campus (fraternity) housing, rented for them as part of the Michigan Group Study Project supervised by Theodore Newcomb from 1953 to 1956. All were incoming transfer students with no prior acquaintance of one another. See $ProjectInfo for more information.

Usage


Newcomb_Fraternity

Format

An object of class list of length 18.

Source

UCINET

References

See $ProjectInfo$References

Examples


Newcomb_Fraternity


[Package xUCINET version 0.0.1.0022 Index]