Rehnquist_SupremeCourt {xUCINET}	R Documentation

Voting among US Supreme Court Justices

Description

This data contains the voting behavior among Supreme Court Justices of the United States between 1995 and 2004. From 1986 to 2005 the chief justice was William Hubbs Rehnquist. The data consist of a 2-mode matrix in which the rows are court cases and the columns are judges. A cell (i,j) = 1 if judge j voted with the majority on case i. See $ProjectInfo for more information.

Usage


Rehnquist_SupremeCourt

Format

An object of class list of length 4.

Source

UCINET

References

See $ProjectInfo$References

Examples


Rehnquist_SupremeCourt


[Package xUCINET version 0.0.1.0022 Index]