Sampson_Monastery {xUCINET}	R Documentation

Sampson's network among monks

Description

The data stem from an ethnographic study of community structure in a New England monastery by Samuel F. Sampson. The study describes several social relations among a group of men (novices) who were preparing to join a monastic order. Some novices had attended the minor seminary of 'Cloisterville' before they came to the monastery. The data sets presented here contain the affect relations among the novices, which were collected by asking them to indicate whom they liked most and whom they liked least. The novices were asked for a first (3), second (2), and third (1) choice on both questions. The social relations were measured at five moments in time. The fourth measurement (at time T4) took place one week before four of the novitiates, Gregory (10), Basil (16), Elias (17) and Simplicius (18), were expelled from the monastery. and the voluntary departure of several others - most immediately, John (9), Hugh (11), Boniface (12), Mark (13) and Albert (14). In the end, only Bonaventure (2), Berthold (4), Ambrose (3) and Louis (6) remained. Most of the present data are retrospective, collected after the breakup occurred. The exceptions are 'LikeT3' data gathered at three times. Based on his observations and analyses, Sampson divided the novices into four groups: Young Turks, Loyal Opposition, Outcasts, and an interstitial group. The Loyal Opposition Bonaventure, Ambrose, Berthold, Peter, Louis consists of the novices who entered the monastery first. The Young Turks arrived later, in a period of change Winfrid, John-Bosco, Gregory, Hugh, Boniface, Mark, Albert. They questioned practices in the monastery, which the members of the Loyal Opposition defended. Some novices did not take sides in this debate, so they are labeled 'interstitial' Romuald, Victor, Amand. The Outcasts are novices who were not accepted in the group Basil, Elias, Simplicius. Information about the senior monks was not included. See $ProjectInfo for more information.

Usage


Sampson_Monastery

Format

An object of class list of length 13.

Source

UCINET

References

See $ProjectInfo$References

Examples


Sampson_Monastery


[Package xUCINET version 0.0.1.0022 Index]