Savage_TransactionFlows {xUCINET}	R Documentation

Savage's transaction flows between countries

Description

This data contain data about transaction flows between 15 countries prior to the economic crash of 1929, as reported by Savage and Deutsch in their paper in 1960. See $ProjectInfo for more information.

Usage


Savage_TransactionFlows

Format

An object of class list of length 4.

Source

UCINET

References

See $ProjectInfo$References

Examples


Savage_TransactionFlows


[Package xUCINET version 0.0.1.0022 Index]