Schwimmer_PapuanVillage {xUCINET}	R Documentation

Schwimmer's gift-giving network

Description

These data represent the relation of gift-giving (taro exchange) among 22 households in a Papuan village. Hage & Harary (1983) used them to illustrate a graph Hamiltonian cycle. Schwimmer points out how these ties function to define the appropriate persons to mediate the act of asking for or receiving assistance among group members. This dataset is called 'Taro' in Windows Ucinet. See $ProjectInfo for more information.

Usage


Schwimmer_PapuanVillage

Format

An object of class list of length 4.

Source

UCINET

References

See $ProjectInfo$References

Examples


Schwimmer_PapuanVillage


[Package xUCINET version 0.0.1.0022 Index]