Wolfe_Primates {xUCINET}	R Documentation

Wolfe's study on interaction among monkeys

Description

These data represent 3 months of interactions among a troop of monkeys, observed in the wild by Linda Wolfe as they sported by a river in Ocala, Florida.

Usage


Wolfe_Primates

Format

A list object (xUCINET project/dataset) containing 2 networks and 3 attributes

$ProjectInfo

- Contains a series of elements and descriptions: $GeneralDescription, $AttributesDescription, $NetworksDescription, $References

$Attributes

- A data.frame with attributes for nodes

$NetworkInfo

- A data.frame with additional information about each network

$Strength

- Network1. See $ProjectInfo$NetworksDescription for more information

$Connection

- Network2. See $ProjectInfo$NetworksDescription for more information

Source

UCINET

References

See $ProjectInfo$References

Examples


Wolfe_Primates


[Package xUCINET version 0.0.1.0022 Index]