Zachary_KarateClub {xUCINET}	R Documentation

Zachary's University Karate Club

Description

Network data among 34 members of a university karate club collected by Wayne Zachary between 1970 and 1972. Zachary (1977) used these data and an information flow model of network conflict resolution to explain the split-up of this group following disputes among the members. See $ProjectInfo for more information.

Usage


Zachary_KarateClub

Format

An object of class list of length 5.

Source

UCINET

References

See $ProjectInfo$References

Examples


Zachary_KarateClub


[Package xUCINET version 0.0.1.0022 Index]