xAddToProject {xUCINET}	R Documentation

Add an additional network to an existing "dataproject"

Description

This function adds a network and information about that network to an existing "dataproject".

Last updated: 28 April 2022.

Usage


xAddToProject(
  ProjectName,
  NetworkName = NULL,
  NETFILE1 = NULL,
  FileType = NULL,
  InFormatType = "AdjMat",
  NetworkDescription = NULL,
  Mode = NA,
  Directed = TRUE,
  Loops = FALSE,
  Values = NA,
  Class = "matrix"
)

Arguments

ProjectName
The name of the "dataproject", to which the network needs to be added.

NetworkName
The name to be given to the new network.

NETFILE1
A single social network file. This could be a csv file or an R object. The input type is defined in "FileType", while the format of the input data is defined in "InFormatType".

FileType
The type of data that is being imported. Default is "csv". Other options are "Robject" when the data is already an object in R, and "csv2". Note that by default a csv file use a point (".") for the decimal point and a comma for the separator in most countries. However, in some Western European countries a comma is used for the decimal point and a semicolon for the separator, which will require "csv2".

InFormatType
The format used to store the network data in a csv file. Currently only reads in an adjacency matrix "AdjMat".

NetworkDescription
Detailed information about the network being imported.

Mode
Character (string), indicating a name for the mode of the network. For two-mode networks this would need to contain the names for both modes. If it is a one-mode network, and Mode=NA then the mode name "A" will be assumed.

Directed
Logical, whether the network should be considered directed or not. The default is TRUE.

Loops
Logical, whether the network's self-nominations (loops) should be considered. By default this is FALSE and loops are not considered.

Values
Character string, indicating what type of values the network contains (e.g., "Binary","Nominal","Ordinal","Rank","Interval","Ratio").

Class
The way the data is being stored. Currently the network will be stored as an adjacency matrix in an object of class "matrix".

Value

A "dataproject" object containing network data.

Author(s)

Filip Agneessens, filipagneessens2@gmail.com

Examples


## We can consider an existing network and add the transposed
## Create a network with ordinal categories (0='no relation', 1='acquaintance', 2='friend',
##    3='good friend') as an (adjacency) matrix:
Friendship3<-matrix(c(0,1,1, 2,0,3, 0,2,0),3,3, byrow=TRUE)
## Add node names to the matrix:
rownames(Friendship3)<-c("a","b","c")
colnames(Friendship3)<-rownames(Friendship3)

## Now create the "dataproject", which we call "SchoolClass_Project1"
SchoolClass_Project1<-xCreateProject(GeneralDescription="Dataset among 3 students",
     NetworkName="Friendship",
     NETFILE1=Friendship3,
     FileType="Robject",
     InFormatType="AdjMat",
     NetworkDescription="Strength (0=no relation, 1=acquaintance, 2=friend, 3=good friend)",
     Mode=c("Students"),
     Directed=TRUE,
     Loops=FALSE,
     Values="Ordinal",
     Class="matrix",
     References="No references")
## Check the newly created "dataproject" object
SchoolClass_Project1

## Now create the transposed
SchoolClass_Project1<-xAddToProject(ProjectName=SchoolClass_Project1,
        NetworkName="Friendship_Tr",
        NETFILE1=t(SchoolClass_Project1$Friendship),
        FileType="Robject",
        InFormatType="AdjMat",
        NetworkDescription="Transposed of the friendship network",
        Mode=c("Students"),
        Directed=TRUE,
        Loops=FALSE,
        Values="Ordinal",
        Class="matrix")
## Check the new result:
SchoolClass_Project1

[Package xUCINET version 0.0.1.0022 Index]