xAttributeToNetwork {xUCINET}	R Documentation

Converts an attribute into a network

Description

This function converts an attribute vector into a network.

Usage


xAttributeToNetwork(ATT1, Type = "Same")

Arguments

ATT1
The attribute vector to be used.

Type
The conversion to be used. The options are: "Same" (gives 1 for ij if i and j have the same value), "AbsDiff" (where ij is the absolute difference in value between i and j), "Diffij" (where ij is the difference: i - j), "Diffji" (where ij is the difference: j - i), "Sender" (where ij is the value for i), "Received" (where ij is the value for j).

Value

An network.

Examples


A1<-c(1,3,1,2,2,3,1,3)
xAttributeToNetwork(A1, Type="Same")

[Package xUCINET version 0.0.1.0022 Index]