xBetaCentrality {xUCINET}	R Documentation

Calculates Beta centrality (Bonacich's power centrality)

Description

Calculates the Beta centrality for a one-mode network, where Beta can be a vector of values or a single value, but must be smaller than 1/highest eigenvalue. Can to be 'problematic' for directed networks. Values can be binary or valued. For valued networks. Different normalizations are available in Option. No missing data are allowed.

Last updated: 1 February 2022.

Usage


xBetaCentrality(NET1, Beta = 0, Loops = FALSE, Option = "None")

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

Beta
A single value or vector contain Beta values which need to be smaller than 1/highest eigenvalue.

Loops
Logical statement, whether Loops should be included.

Option
The normalization chosen, which can be one of the following "None" (no normalization), "MAX1" (set maximum value to 1), "SUM1" (set sum of values to 1), "SSQ1" (set sum of squares equal to 1), "SSQN" (set sum of squares equal to number of nodes)

Value

A matrix with Beta centrality values.

Examples


## Undirected one-mode network
xBetaCentrality(ASNR_Fig09x3, c(0,.1,.2,.3,.4,-.1,-.2), Option="SSQN")

## Directed one-mode network
xBetaCentrality(ASNR_Fig09x4, Beta=c(0,1,10,100,1000) , Option="SSQN")
xBetaCentrality(t(ASNR_Fig09x4), Beta=c(0,1,10,100,1000) , Option="SSQN")


[Package xUCINET version 0.0.1.0022 Index]