xBetaReachBetweennessCentrality {xUCINET}	R Documentation

Calculate Beta Reach betweenness centrality.

Description

Calculates the Beta Reach between centrality for a one-mode or two-mode network, i.e., the number (proportion) of times a node is in between others, where this is weights by its length. Beta needs to be a single value (default=.5) Missing data should be indicated by 'NA'.

Last updated: 1 February 2022.

Usage


xBetaReachBetweennessCentrality(NET1, Beta = 0.5)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

Beta
A single value for Beta (max = 1).

Value

A matrix with Beta reach betweenness centrality values.

Examples


## Undirected one-mode network
xBetaReachBetweennessCentrality(Padgett_FlorentineFamilies$Marriage, Beta=.5)
## Directed one-mode network
xBetaReachBetweennessCentrality(Krackhardt_HighTech$Advice, Beta=.3)


[Package xUCINET version 0.0.1.0022 Index]