xBetaReachCentrality {xUCINET}	R Documentation

Calculate Beta Reach centrality.

Description

Calculates the Beta Reach centrality for a one-mode or two-mode network, i.e., the number/proportion of nodes reachable in k steps weighted by Beta. Beta can be a single value or set of values.(default=.5) The Loops for the geodesic distance matrix is by default set to NA. Missing data should be indicated by 'NA'.

Last updated: 1 February 2022.

Usage


xBetaReachCentrality(NET1, Beta = 0.5, Loops = FALSE, Normalize = FALSE)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

Beta
A single value or vector contain Beta values.

Loops
Logical statement, whether Loops should be included (default=FALSE).

Normalize
Whether to return the normalized value (divide by theoretical maximum value) or raw counts.

Value

A matrix with Beta reach centrality values.

Examples


## Undirected one-mode network
xBetaReachCentrality(Padgett_FlorentineFamilies$Marriage, Beta=c(.5,.3))
## Directed one-mode network (out-beta reach followed by in-beta reach)
xBetaReachCentrality((Krackhardt_HighTech$Advice), Beta=c(.4), Normalize=TRUE)
xBetaReachCentrality(t(Krackhardt_HighTech$Advice), Beta=c(.4), Normalize=TRUE)


[Package xUCINET version 0.0.1.0022 Index]