xBiCliques {xUCINET}	R Documentation

Calculates the cliquemembership for a two-mode network.

Description

This function takes a two-mode network stored as a matrix and calculates the bicliques. By default a biclique consists of at least 3 mode A and 3 mode B nodes who are all maximal connected.

Usage


xBiCliques(NET2, MinA = 3, MinB = 3)

Arguments

NET2
A single two-mode network stored in an R object as a matrix.

MinA
The minimum number of mode A nodes (row)

MinB
The minimum number of mode B nodes (columns)

Value

A one-mode bipartite network

Examples


xBiCliques(Davis_SouthernWomen$Attendance)


[Package xUCINET version 0.0.1.0022 Index]