xBlockmodel {xUCINET}	R Documentation

Construct a blockmodel

Description

Takes a network and an attribute file containing partitions.

Usage


xBlockmodel(NET1, ATTR1)

Arguments

NET1
A network

ATTR1
An attribute file containing partition to be performed.

Value

plot

References

Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE. Doreian, P., Batagelj, V., & Ferligoj, A. (2005). Generalized blockmodeling, (Structural analysis in the social sciences, 25). Cambridge: Cambridge University Press. Žiberna, A. (2007). Generalized Blockmodeling of Valued Networks. Social Networks, 29(1), 105-126. doi: 10.1016/j.socnet.2006.04.002 Žiberna, A. (2008). Direct and indirect approaches to blockmodeling of valued networks in terms of regular equivalence. Journal of Mathematical Sociology, 32(1), 57-84. doi: 10.1080/00222500701790207 Žiberna, A. (2014). Blockmodeling of multilevel networks. Social Networks, 39(1), 46-61. doi: 10.1016/j.socnet.2014.04.002

Examples


NET1<-matrix(c(0,1,0,1,0,
               0,0,0,0,1,
               1,1,0,1,1,
               0,0,0,0,1,
               1,1,1,0,0),5,5)
xBlockmodel(NET1,ATTR1=c(1,2,1,2,1))

[Package xUCINET version 0.0.1.0022 Index]