xCentralization {xUCINET}	R Documentation

Calculates the (out)degree centralization for a one-mode or two-mode network.

Description

Using a one-mode or two-mode binary network, this function calculates the level of centralization in a network based on (out)degree. It takes the maximum observed (out)degree as a point of reference against which the (out)degrees for all other nodes are compared. No missing data allowed. The Loops is ignored.

Usage


xCentralization(NET1)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

Value

A single value.

References

Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

See Also

xDensity(), xDegreeCentrality()

Examples


## Two simple undirected one-mode networks (with the same value for connectedness)
xProportionReach(ASNR_Fig10x3C, kReach=2)
xProportionReach(ASNR_Fig10x3D, kReach=3)
## Adding an attribute for rows and columns:
xProportionReach(ASNR_Fig10x3D,ROWS=c(1,1,1,2,2,2),COLS=c(1,1,1,2,2,2))

## Some more complex examples with a two-mode network
NET2<-matrix(c(0,1,0,0,0,0,0, 1,0,0,0,0,0,0, 1,1,1,0,0,1,0, 0,1,1,0,0,0,0,
               0,0,0,1,1,0,0, 1,0,0,0,0,1,0, 0,1,1,1,0,1,0, 0,0,1,0,0,0,0),8,7,byrow=TRUE)
## When running the function on the twomode, note the use of "A" and "B" to separate
## mode A (row) from mode B (column) nodes:
xProportionReach(NET2)
## When adding an attribute for ROWS, note that mode A nodes are now split further
## into 2 groups: A_1, A_2 and A_3.
xProportionReach(NET2,ROWS=c(1,1,1,2,2,2,3,3))
## We can also keep the values separate for each unique node in the column:
xProportionReach(NET2,ROWS=c(1,1,1,2,2,2,3,3),COLS="Unique")
## Or have a different attribute for mode A and mode B:
xProportionReach(NET2,ROWS=c(1,1,1,2,2,2,3,3),COLS=c("P","E","T","T","P","T","E"))
## Or combine unique (in the rows) with aggregating all (in the columns):
xProportionReach(NET2,ROWS="Unique")


[Package xUCINET version 0.0.1.0022 Index]