xCliquesMembership {xUCINET}	R Documentation

#——————————————————————– Provides the clique membership for a one-mode network

Description

Using an binary, undirected one-mode network (NET1), this function gives the maximum clique solutions. Clique membership is captured by a matrix, with nodes as rows and each clique as a column. The minimum size of the cliques being considered are defined in (Min). The network should be a binary and undirected one-mode network.

Usage


xCliquesMembership(NET1, Min = 3)

Arguments

NET1
A binary, undirected one-mode network stored as a 'matrix' object.

Min
The minimum size for cliques being reported. By default a clique has at least 3 nodes.

Value

A matrix with membership for each node (row) by clique (column).

References

Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE. igraph package - https://igraph.org

See Also

xCliquesCoMembership(), xCliquesOverlap(), igraph::max_cliques()

Examples


## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
xCliquesMembership(Hawthorne_BankWiring$Game)

[Package xUCINET version 0.0.1.0022 Index]