xCliquesOverlap {xUCINET}	R Documentation

Provides the cliques-overlap based on a one-mode network

Description

Using an binary, undirected one-mode network (NET1), this function calculates all the maximum cliques in a network and then defines for the each pair of cliques how many nodes they have in common. The network should be a binary and undirected one-mode network.

Usage


xCliquesOverlap(NET1, Min = 3)

Arguments

NET1
A binary, undirected one-mode network stored as a 'matrix' object.

Min
The minimum size for cliques being considered. By default a clique has at least 3 nodes.

Value

A matrix where each row and column are unique cliques found, and the values are the number of nodes two cliques have in common.

References

Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE. igraph package - https://igraph.org

See Also

xCliquesMembership(), xCliquesMembership(), igraph::max_cliques()

Examples


## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
xCliquesOverlap(Hawthorne_BankWiring$Game)

[Package xUCINET version 0.0.1.0022 Index]