xCombineTies {xUCINET}	R Documentation

Combine two networks

Description

This function allows different forms of combining two networks.

Usage


xCombineTies(NET1, NET2, Type = "Max")

Arguments

NET1
The first network to be used.

NET2
The second network to be used.

Type
The way to combine two matrices. The six options are: "Min" (take the minimum of both ties), "Max" (take the maximum of both ties - the default), "Av" (take the average of both ties), "Sum" (take the sum of both ties), "Prod" (take the product of both ties), "Unique" (creates a unique code for both - only to be used for matrices with limited number of values).

Value

The combined network.

Examples


M1<-matrix(c(1,2,3,4,5,6,7,8,NA),3,3)
M2<-matrix(c(3,2,1,4,5,6,7,8,NA),3,3)
xCombineTies(M1,M2)

[Package xUCINET version 0.0.1.0022 Index]