xComponents {xUCINET}	R Documentation

Calculates information related to the number of components in a network.

Description

Identifies the component for a one-mode or two-mode binary network and provides additional information. It provides information about the number of components, the size of the largest (maximum) and smallest (minimum) component, the mean and median size of the different components. It also calculates the component ratio, which is the largest component minus one divided by n-1. The minimum component size to be considered can be defined by (MinCompSize). The default is to set it to value 1, so isolates are also considered in the component count. For a two-mode network, separate information is provided about both modes. This function relies on the (component.dist) function in the sna package, and for directed networks this offers a choice between ("strong","weak","unilateral", and "recursive"). Missing data (indicated by NA) are not allowed.

Usage


xComponents(
  NET1,
  Measures = c("Min", "Max", "ComponentRatio", "Mean", "Median"),
  MinCompSize = 1,
  Type = "strong"
)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

Measures
Optional measures to be calculated. These include providing information about the size of the smallest component ("Min"), the largest component ("Max"), the component ratio ("ComponentRatio"), the average component size over all components ("Mean"), and the median ("Median").

MinCompSize
The minimum size of the components to be considered. Components smaller than this will be ignored from the analysis. (Default is to consider all, including isolates, i.e., components of size 1).

Type
Whether to consider a weak, strong (or other) way to identify components. This relies on the component.dist function in the sna package, and feeds it directly into the option "connected" in the component.dist function.

Value

A matrix containing multiple measures (rows) and a single column (for a one-mode matrix) or 3 columns ("ModeA", "ModeB" and "Both", for a two-mode network).

References

Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE. Carter T. Butts (2020). sna: Tools for Social Network Analysis. R package version 2.6. http://CRAN.R-project.org/package=sna

See Also

sna::component.dist(), xConnectedness()

Examples of undirected one-mode networks (see Borgatti et al., 2022, Figure 10.3)

xComponents(ASNR_Fig10x3A) xComponents(ASNR_Fig10x3B) xComponents(ASNR_Fig10x3C) xComponents(ASNR_Fig10x3D) xComponents(ASNR_Fig10x3E)

An examples of a directed one-mode networks

NET1<-matrix(c(0,1,1,0,0,0,0,0, 1,0,1,1,0,0,0,0, 1,1,1,0,0,0,0,0, 0,1,1,0,0,0,0,0, 0,0,0,0,0,0,0,0, 0,0,0,0,0,0,1,0, 0,0,0,0,0,1,0,1, 0,0,0,0,0,0,0,0),8,8,byrow=TRUE) xComponents(NET1, Type="weak") xComponents(NET1, Type="strong") xComponents(NET1, Type="strong",MinCompSize = 2)

A two-mode network

NET2<-matrix(c(0,1,0,0,0,0,0, 1,0,0,0,0,0,0, 1,1,1,0,0,0,0, 0,1,1,0,0,0,0, 0,0,0,1,1,0,0, 0,0,0,0,0,1,0, 0,0,0,0,0,1,0, 0,0,0,0,0,0,0),8,7,byrow=TRUE) xComponents(NET2) xComponents(NET2, MinCompSize = 2) xComponents(NET2, MinCompSize = 3)

[Package xUCINET version 0.0.1.0022 Index]