xConnectedness {xUCINET}	R Documentation

Calculates the connectedness (fragmentation) of a network or for subparts of a network.

Description

Using a one-mode or two-mode binary network, this function calculates the connectedness (or its inverse the fragmentation of a network). Connectedness is the proportion of pairs of nodes that can reach each other, irrespective of how long it takes, which is equivalent to the proportion of pairs of nodes that are located in the same component. In practice it takes the reachability/connectedness matrix where a cell has value 1 if i and j can reach each other and takes the average (Option="Mean") or the sum (Option="Sum") of pairs of nodes that can reach each other in a network.

For a two-mode network, the result can be split into three parts (within mode A, within mode B, and between mode A and mode B nodes), and therefore we will obtain a matrix with within and between values.

Missing data NA are not allowed.

In addition, if an attribute file is specified for rows (ROWS) and/or the columns (COLS), then the average or sum is provided for each unique group in that attribute file. This might be useful to examine whether specific types of actors tend to be reachable from specific other nodes. Note that ROWS and COLS does not have to contain the same attribute.

If ROWS or COLS is set to "Unique" (default is "All"), then the matrix is not aggregated for that row or column, and so reachability is obtained for each node. This might be useful if you want to know, for example, if specific nodes are reachable from nodes with specific types of actors.

Usage


xConnectedness(NET1, ROWS = "All", COLS = "All", Option = "Mean")

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

ROWS
Whether to calculate the sum or average of connectedness over all nodes in the row (="All"), whether to keep each node in the row as a unique entity ("Unique") or whether to aggregate based on some attribute variable (in which case (ROWS= name of the attribute dataset).

COLS
Same as for rows, but here for columns. Note that ROWS and COLS can contain a different attribute variable, for example to see whether there is any difference in male or females are able to reach more high achieving versus low achieving students (here gender would be ROWS and academic achievement would be COLS).

Option
Whether to obtain the average ("Mean") or sum ("Sum") for the connectedness matrix.

Value

A single value or matrix calculating the sum or average of reachability across all nodes, or split by category (or node).

References

Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE. Carter T. Butts (2020). sna: Tools for Social Network Analysis. R package version 2.6. http://CRAN.R-project.org/package=sna

See Also

xDensity(), xCompactness(), sna::geodist()

Examples


## A simple undirected one-mode network
xConnectedness(ASNR_Fig10x3D)
## Adding an attribute for rows and columns:
xConnectedness(ASNR_Fig10x3D,ROWS=c(1,1,1,2,2,2),COLS=c(1,1,1,2,2,2))

## Some more complex examples with a two-mode network
NET2<-matrix(c(0,1,0,0,0,0,0,
              1,0,0,0,0,0,0,
              1,1,1,0,0,0,0,
              0,1,1,0,0,0,0,
              0,0,0,1,1,0,0,
              0,0,0,0,0,1,0,
              0,0,0,0,0,1,0,
              0,0,0,0,0,0,0),8,7,byrow=TRUE)
## When running the function on the twomode, note the use of "A" and "B" to separate
## mode A (row) from mode B (column) nodes:
xConnectedness(NET2)
## When adding an attribute for ROWS, note that mode A nodes are now split further
## into 2 groups: A_1, A_2 and A_3.
xConnectedness(NET2,ROWS=c(1,1,1,2,2,2,3,3))
## We can also keep the values separate for each unique node in the column:
xConnectedness(NET2,ROWS=c(1,1,1,2,2,2,3,3),COLS="Unique")
## Or have a different attribute for mode A and mode B:
xConnectedness(NET2,ROWS=c(1,1,1,2,2,2,3,3),COLS=c("P","E","T","T","P","T","E"))
## Or combine unique (in the rows) with aggregating all (in the columns):
xConnectedness(NET2,ROWS="Unique")


[Package xUCINET version 0.0.1.0022 Index]