xCorePeriphery {xUCINET}	R Documentation

Finds a core-periphery structure

Description

Takes a network and finds the core-periphery structure using the correlation between the values and a vector with 0s and 1s (concentration).

Usage


xCorePeriphery(NET1)

Arguments

NET1
A network stored as a matrix

Value

The eigenvector of each node, the correlation and an optimal solution.

References

Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

Examples


NET1<-matrix(c(0,1,0,1,0,
               0,0,0,0,1,
               1,1,0,1,1,
               0,0,0,0,1,
               1,1,1,0,0),5,5)
xCorePeriphery(NET1)

[Package xUCINET version 0.0.1.0022 Index]