xCorrelation {xUCINET}	R Documentation

Performs a correlation test using a classic (and a permutation-based) significance test between two vector variables

Description

This function takes two variables (vectors of equal length) and returns the correlation coefficient("CorrCoef") and the classic (two-tailed) significance test ("ClassicSign_2tailed"). If NPERM (number of permutations) is set to an integer value higher than 0 (default is 1000), a permutations-based significance is also provided, which will provide information about the correlations coefficients for the permuted variables (including the histogram). A higher value than 1000 is generally advisable. Both the one-tailed ("AsSmall" or "AsLarge") and the two-tailed ("AsExtreme") test are given. Note that cases with missing values are allowed, but that this can impact the degrees of freedom for the correlations obtained for the permutations. For example, suppose the same 20% of cases are missing for both variables (i.e., unit non-response), then permuting these missing values will mean that a lot more cases will have a missing for (at least) one of both variables.

Last updated: 26 April 2022.

Usage


xCorrelation(VEC1, VEC2, NPERM = 1000)

Arguments

VEC1
A first vector of values.

VEC2
A second vector of values, of equal length as VEC1.

NPERM
A positive integer value indicating the number of permutations to be used for the statistical test. If NPERM is set to NULL, NA or 0, no permutation-based significance test is performed.

Value

Returns an object containing the correlation coefficient ($CorrCoeff) and standard significance test (ClassicSign_2tailed), and in the case a permutation test is performed, in addition a significance based on the permutation ("AsSmall", "AsLarge", and "AsExtreme").

Author(s)

Filip Agneessens filipagneessens2@gmail.com

Examples


RESULS1<-xCorrelation(c(2,4,2,3,3,1),c(4,2,3,2,1,5), NPERM=5000)
RESULS1
RESULS2<-xCorrelation(c(2,3,2,1,2,3,2,1),c(4,2,3,2,4,2,3,2), NPERM=5000)
RESULS2
xCorrelation(c(1,4,2,1),c(4,2,3,3), NPERM=0)
xCorrelation(c(1,4,2,1),c(4,2,3,3), NPERM=NA)
xCorrelation(c(1,4,2,1),c(4,2,3,3), NPERM=NULL)

[Package xUCINET version 0.0.1.0022 Index]