xCorrespondenceAnalysis {xUCINET}	R Documentation

Performs a correspondence analysis (CA) on a matrix

Description

Takes a rectangular matrix and performs a correspondence analysis on the matrix. This function relies on the corresp() function (MASS).

Usage


xCorrespondenceAnalysis(
  MAT1,
  Margins = 0.05,
  Distance = c(0, 4),
  DIM = 2,
  Color = c("Blue", "DarkRed")
)

Arguments

MAT1
A rectangular matrix

Margins
The margins added to the output figure

Distance
The distance of the names (labels) of the nodes to the midpoint (as a % of the overall distance).

DIM
The number of dimensions to be considered

Color
A vector of size 2, indicating the colors for row and column variables.

Value

A figure containing the correspondence analysis and the output from corresp().

References

Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

Examples


MATY<-matrix(c(6,3,3,2,
               8,2,5,2,
               4,8,4,5,
               9,8,6,3,
               7,3,2,5,
               3,2,1,5,
               8,5,6,7,
               3,2,4,2),8,4,byrow=TRUE)
rownames(MATY)<-c("P1","P2","P3","P4","P5","P6","P7","P8")
colnames(MATY)<-c("V1","V2","V3","V4")
xCorrespondenceAnalysis(MATY)

[Package xUCINET version 0.0.1.0022 Index]