xCyclicality {xUCINET}	R Documentation

Calculates the level of cyclicality in a binary, one-mode network.

Description

This function takes a one-mode network and provides information related to the level of cyclicality in the network. For an undirected network this is also sometimes referred to as clustering (and will give the same result as transitivity). The basis for calculating the level of cyclicality is the number of cyclical triplets (where i->j=1, j->k=1 and k->i=1) divided by the number of opportunities for a cyclical triplet (where i->j=1 and j->k=1, but where k->i could be either 0 or 1). Note that the Loops entries (self-nominations) are always ignored for this function and that two-mode data are not allowed. Triplets where information on any of the three cells (i,j), (j,k) or (k,i) are missing (indicated by values 'NA' in the matrix) are ignored. The output generates:

Number and proportion of valid triplets, i.e., where (i,j), (j,k) and (k,i) are either 0 or 1, but not NA.

The density for the network, i.e. the proportion of ties with values 1 (considering only those triplets that are valid),

The number of triplets belonging to any of the unique 8 different types of triplets,

The Cyclicality Index: (ij=1 & jk=1 & ki=1)/((ij=1 & jk=1 & ki=1)+(ij=1 & jk=1 & ki=0)), which can be compared to the density.

Usage


xCyclicality(NET1)

Arguments

NET1
A one-mode network stored as a 'matrix' object.

Value

A vector with a series of measures. See Description.

References

Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

See Also

xDensity(), xReciprocity(), xTransitivity()

Examples


## A network with 6 valid triplets out of 4*3*2 = 0.25, where the density is 0.833
## and the index is 3/(3+1)=.75.
NET1<-matrix(c(0,1,NA,0,
      1,0,1,NA,
      1,0,1,1,
      0,NA,1,0),4,4,byrow=TRUE)
xCyclicality(NET1)

## Cyclicality for Krackhardt's high-tech managers (index of 0.38, 0.28 and 0.00 respectively):
xTransitivity(Krackhardt_HighTech$Advice)
xTransitivity(Krackhardt_HighTech$Friendship)
xTransitivity(Krackhardt_HighTech$ReportTo)

## To show that the cyclicality index would be similar to the density for a
## random network, let's generate 100 networks of size 50 by 50 with density .3:
## Store the results for each generated network in RES1:
RES1<-matrix(NA,100,12)
## Now to generate 100 networks use a loop:
for (k1 in 1:100)
  {
  MA1<-matrix(runif(2500)>=.7,50,50)
  RES1[k1,]<-t(xCyclicality(MA1))
  }
## We can check that the density for the generated networks is around .3:
mean(RES1[,3])
## Now, we can check that the cyclicality index for the generated graphs.
## This is indeed around .3:
mean(RES1[,12])
## Proving that the cyclicality index will be close to the density for a random graph.

[Package xUCINET version 0.0.1.0022 Index]