xDichotomize {xUCINET}	R Documentation

Dichotomizes the ties of a network

Description

This function offers different options for dichotomizing a network. The default is "GT" with Value=0.

Usage


xDichotomize(NET1, Value = 0, Type = "GT")

Arguments

NET1
A network to be dichotomized.

Value
The value at which to dichotomize. When using (Type="EQ") this can also be a vector of values to be set to 1.

Type
The type of dichotomization. The five options are: "GT" (greater than), "LT" (less than), "EQ" (equal),

Value

A dichotomized network.

Examples


M1<-matrix(c(1,2,3,4,5,6,7,8,NA),3,3)
xDichotomize(M1,Value=c(4,7,2,20),Type="EQ")
xDichotomize(M1,Value=c(4),Type="GT")

[Package xUCINET version 0.0.1.0022 Index]