xDualCorePeriphery {xUCINET}	R Documentation

Performs a dual core-periphery analysis.

Description

Performs a separate core-periphery analysis on row and column nodes using one-mode projections.

Usage


xDualCorePeriphery(NET2)

Arguments

NET2
A single two-mode network stored in an R object as a matrix.

Value

A membership

Examples


xDualCorePeriphery(Davis_SouthernWomen$Attendance)

[Package xUCINET version 0.0.1.0022 Index]