xEgoAlterSimilarityCat {xUCINET}	R Documentation

Information based the similarity/difference on a categorical between ego and the alters that ego is directly connected with.

Description

This function takes a one-mode network (NET1) and a vector containing a categorical nodal attribute (ATTR1). It returns for each node in the rows of the matrix information related to the similarity or difference in attribute data between the node and its alters. As an output it provides by default: -"a": the number of connected alters that belong to the same category as ego. -"b": the number of connected alters that belong to a different category than ego. -"c": the number of alters that ego is not connected to, that belong to the same category as ego. -"b": the number of alters that ego is not connected to, that belong to a different category than ego.

Usage


xEgoAlterSimilarityCat(
  NET1,
  ATTR1,
  Measures = c("PctSame", "EIIndex", "OddsRatio", "LogOddsRatio", "YulesQ"),
  Loops = TRUE
)

Arguments

NET1
A network dataset stored as an object of class "matrix". A one-mode network with (n) nodes, or a two-mode network with (m) nodes of mode "A" (rows) and (n) nodes of mode "B" (columns).

ATTR1
A continuous attribute vector of size (n).

Measures
The choice of output to be provided. See description.

Loops
Whether to ignore the Loops (default is to ignore the Loops).

Value

A matrix containing different measures (columns) for each node (row).

Examples


## Consider the Friendship network for Krackhardt's high-tech managers and the attribute age.
xEgoAlterSimilarityCat(Krackhardt_HighTech$Friendship,Krackhardt_HighTech$Attributes$Department)

[Package xUCINET version 0.0.1.0022 Index]