xEgoAlterSimilarityCon {xUCINET}	R Documentation

Information based the similarity in continuous characteristics between ego and the alters that ego is directly connected with.

Description

This function takes a one-mode network (NET1) and a vector containing a continuous nodal attribute (ATTR1). It returns for each node in the rows of the matrix information related to the similarity or difference in attribute data between the node and its alters. For valued networks it weights the impact of alters by the strength of its ties. As an output it provides by default: -"AvDiffAv": (Weighted) average of difference between alter's values and ego's own value. -"AvAbsDiff": (Weighted) average of the absolute difference between alter's values and ego's own value. -"AvSqDiff": (Weighted) average of the squared differences between alter's values and ego's own value. -"Corr_AbsDiff": For any ego, correlation between ego's (valued) link to alter and the absolute difference in value between ego and alter. Has n-1 alters as cases. -"Corr_Zegers": For any ego, correlation between ego's link to alter and the product of values for ego and alter divided by their squared values. -"Corr_MinOverMax": For any ego, correlation between ego's link to alter and the lowest over the highest value. -"Corr_Product": For any ego, correlation between ego's link to alter and the product of values for ego and alter.

Usage


xEgoAlterSimilarityCon(
  NET1,
  ATTR1,
  Measures = c("AvDiff", "AvAbsDiff", "AvSqDiff", "Corr_AbsDiff", "Corr_DiffSq",
    "Corr_Zegers", "Corr_MinOverMax", "Corr_Product"),
  Loops = FALSE
)

Arguments

NET1
A network dataset stored as an object of class "matrix". A one-mode network with (n) nodes, or a two-mode network with (m) nodes of mode "A" (rows) and (n) nodes of mode "B" (columns).

ATTR1
A continuous attribute vector of size (n).

Measures
The choice of output to be provided. See description.

Loops
Whether to ignore the Loops (default is to ignore the Loops).

Value

A matrix containing different measures (columns) for each node (row).

Examples


## Consider the Friendship network for Krackhardt's high-tech managers and the attribute age.
xEgoAlterSimilarityCon(Krackhardt_HighTech$Friendship,Krackhardt_HighTech$Attributes$Age)

[Package xUCINET version 0.0.1.0022 Index]