xEigenvectorCentrality {xUCINET}	R Documentation

Calculates the eigenvector centrality for a network (i.e. for each row).

Description

Calculates the eigenvector centrality for a one-mode network. Tends to be 'problematic' for directed networks. Values can be binary or valued. For valued networks. The normalized version ensures that the maximum value is 1. No missing data are allowed.

Last updated: 1 February 2022.

Usage


xEigenvectorCentrality(NET1, Loops = FALSE)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

Loops
Logical statement, whether Loops should be used.

Value

A matrix with eigenvector centrality and normalized eigenvector centrality (max=1).

Examples


##Undirected networks
xEigenvectorCentrality(ASNR_Fig09x1)
xEigenvectorCentrality(ASNR_Fig09x2)

[Package xUCINET version 0.0.1.0022 Index]