xGroupHomophily {xUCINET}	R Documentation

Calculates measures of homophily for a one-mode network.

Description

Using a one-mode binary network, this function calculates measures of homophily. This includes the proportion homophilous ties, the EI index and Yule's Q. Missing data are allowed. The Loops is ignored. Provides output: -"a": the number of pairs of nodes that have a tie and where both nodes belong to the same category. -"b": the number of pairs of nodes that have a tie and where both nodes belong to a different category. -"c": the number of pairs of nodes that do not have a tie and where both nodes belong to the same category. -"d": the number of pairs of nodes that do not have a tie and where both nodes belong to a different category.

Usage


xGroupHomophily(NET1, ATTR1)

Arguments

NET1
A one-mode binary network stored as a 'matrix' object.

ATTR1
A vector containing the attribute.

Value

A vector with measures related to homophily.

References

Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

See Also

xDensity(), xEgoAlterSimilarityCat()

Examples


## Two examples of directed networks (See chapter 10 of Borgatti et al., 2022):
xGroupHomophily(Krackhardt_HighTech$Advice,Krackhardt_HighTech$Attributes$Department)
xGroupHomophily(Krackhardt_HighTech$Friendship,Krackhardt_HighTech$Attributes$Department)


[Package xUCINET version 0.0.1.0022 Index]