xHierarchicalClustering {xUCINET}	R Documentation

Provides the hierarchical clustering solution for a square matrix

Description

Using a square matrix, representing either similarities or differences, this function calculates the clustering solution. This function relies on hclust and plot, and therefore incorporates their respective options. If requested (by adding values to NOC), cluster membership is offered for a solution with k clusters.

Usage


xHierarchicalClustering(
  NET1,
  Input = NULL,
  Method = "complete",
  TitleDendrogram = "Cluster Dendrogram",
  NOC = NULL
)

Arguments

NET1
A binary, undirected one-mode network stored as a 'matrix' object.

Input
Whether the input matrix contains similarities ("Similarities") or differences ("Differences")

Method
which clustering method to use. Default is "complete". See the hclust function.

TitleDendrogram
The title to be used in the plot

NOC
If containing a single value or vector, cluster membership is provided as output.

Value

A matrix where each row and column are unique cliques found, and the values are the number of nodes two cliques have in common.

References

Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

See Also

xCliquesMembership(), xCliquesMembership(), hclust()

Examples


## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
CCM<-xCliquesCoMembership(ASNR_Fig11x1,Min=3)
xHierarchicalClustering(CCM, Input="Similarities", Method="single")
MATY<-matrix(c(2,3,1,
               3,5,6,
               1,6,4),3,3)
xHierarchicalClustering(MATY, Input="Similarities")

[Package xUCINET version 0.0.1.0022 Index]