xInducedCentrality {xUCINET}	R Documentation

Calculate induced centrality.

Description

Induced centrality calculates the change in specific network properties (graph invariants) when a node is removed. A graph invariant can be used to induce a centrality score by measuring the contribution each node makes to the network property (e.g. the number of components). The induced centrality of a node is the value for the graph invariant for the observed network minus the value for the graph invariant for the observed network but with the node deleted. This routine calculates the induced centrality for the following network properties:

"kReach1": Number of nodes reachable within geodesic distance 1 (Degree)

"kReach2": Number of nodes reachable within geodesic distance 2

"kReach3": Number of nodes reachable within geodesic distance 3

"kReach4": Number of nodes reachable within geodesic distance 4

"kReachAll": Number of nodes reachable within a finite number of nodes

"RecipCloseness": Sum of reciprocal geodesic distance to all other nodes

"Betweenness": Number of times a node is between other nodes No missing data are allowed.

Last updated: 1 February 2022.

Usage


xInducedCentrality(NET1)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

Value

A matrix with a series of induced centrality measures, followed by a split between endogeneous and exogeneous.

Examples


## Undirected one-mode network
xInducedCentrality(Padgett_FlorentineFamilies$Marriage)

## Directed one-mode network
xInducedCentrality(Krackhardt_HighTech$Advice)
xInducedCentrality(t(Krackhardt_HighTech$Advice))

[Package xUCINET version 0.0.1.0022 Index]