xLabelPropagation {xUCINET}	R Documentation

Provides the membership of nodes for the community detection using label propagation.

Description

Using an undirected one-mode network (NET1), this function calculates the solution for community detection membership using label propagation and graphically shows the result. The output is a matrix providing membership for communities with (k) groups. This function relies on igraph's cluster_label_prop() function.

Usage


xLabelPropagation(
  NET1,
  vertex.shape = NULL,
  vertex.label.cex = NULL,
  layout = NULL
)

Arguments

NET1
An undirected one-mode network stored as a 'matrix' object.

vertex.shape
An option argument to specify the shapes of the nodes. Needs to be a vector of length equal to the number of nodes.

vertex.label.cex
The size of the font used for the labels (default when NULL is cex=1).

layout
The layout option used. See igraph. (default when NULL is layout_nicely, which tries to find the optimal layout).

Value

A vector with membership for each node (row).

References

Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE. igraph package - https://igraph.org

See Also

xGirvanNewman(), xFastGreedy(), xWalkTrap(), xLouvainMethod(), igraph::cluster_label_prop()

Examples


## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
xLabelPropagation(Zachary_KarateClub$Connection, vertex.label.cex=.7)
## Add attribute file related to the club split:
## Define the shapes to be used
UseShapes<-c("circle","rectangle")
## Define the vector for nodes -
ZKAttrClub<-(Zachary_KarateClub$Attributes$Club>1)+1
xLabelPropagation(Zachary_KarateClub$Connection, vertex.shape=UseShapes[ZKAttrClub],
         vertex.label.cex=.7)

## A second dataset:
xLabelPropagation(Kapferer_TailorShop$SociationalT1, vertex.label.cex=.7,
                  layout="layout_with_fr")

[Package xUCINET version 0.0.1.0022 Index]