xMultipleTieComposition {xUCINET}	R Documentation

Provides the outdegree for multiple types of network relations among the same nodes and a measure of heterogeneity (IQV)

Description

Takes an object of class 'list', with a set of different types of network relations (matrices) among the same nodes stored as elements. For the different types of networks it provides for each node:

the number of valid alters, i.e., the number of alters for which information is available (not equal to NA),

the degree, i.e., the sum of ties (or tie strengths) to alters (Measures="SumStrength"),

the average number of ties (or tie strengths) to alters relative to the number of valid cases (Measures="AvStrength"),

the proportion of ties (or tie strengths) for a specific relation relative to the total ties over all relations (Measures="PropStrength"),

a measure of heterogeneity using Agresti's index of qualitative variation (IQV) (Measures="IQVType").

Usage


xMultipleTieComposition(
  LIST1,
  Measures = c("SumStrength", "AvStrength", "PropStrength", "IQVType"),
  Loops = FALSE,
  Mode = "OneMode"
)

Arguments

LIST1
A set of matrices stored as a list object

Measures
The measures to be calculated about the set of networks. Options are: "SumStrength","AvStrength","PropStrength" and "IQVType".

Loops
Logical, whether the diagonal values should be ignored (default setting is FALSE, i.e., to ignore the diagonal values)

Mode
Whether the set of networks are one-mode "OneMode" or two-mode networks "TwoMode".

Value

A matrix with different measures (columns) for each node (row)

References

Chapter 8 in Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

Examples


## Consider the different types of networks in the Hawthorne bank wiring room network.
## These are elements 4 to 9 in the project/dataset "Hawthorne_BankWiring".
## All networks are binary, except the last, 9th network ("TradeJobs").
## Although the function can deal with valued networks, we decide to dichotomize the last
## network first and store this as a new (10th) element in the list object:
Hawthorne_BankWiring$TradeJobsDich<-(Hawthorne_BankWiring$TradeJobs>0)*1
## We can now apply the function to all binary networks:
xMultipleTieComposition(Hawthorne_BankWiring[c(4:8,10)])
## See Chapter 8 in Borgatti, Everett, Johnson and Agneessens (2022)

[Package xUCINET version 0.0.1.0022 Index]