xNegativeDegreeCentrality {xUCINET}	R Documentation

Calculates the outdegree for negative relations (i.e. for each row).

Description

Calculates the (out)degree centrality for a one-mode containing a negative relation. To calculate the indegree simply transpose the matrix by using t(). This function calculates degree centrality for negative relations as explained in Everett and Borgatti (2014). Only for binary data. Any missing data (indicated by values 'NA') are ignored.

Last updated: 1 February 2022.

Usage


xNegativeDegreeCentrality(NET1N)

Arguments

NET1N
A one-mode or two-mode negative network stored as a 'matrix' object.

Value

A matrix with outdegree centrality measures.

References

Everett, M.G. Borgatti, S.P. (2014) Networks containing negative ties. Social Networks 38, 111-120.

Examples


## Undirected one-mode network # TABLE 1
xNegativeDegreeCentrality(Read_NewGuinea$Opposition)

## Directed one-mode network # TABLE 3
NEG1<-((Sampson_Monastery$Disesteem+Sampson_Monastery$Dislike+Sampson_Monastery$Blame
        +Sampson_Monastery$NegativeInfluence)>0)*1
xNegativeDegreeCentrality(NEG1)
xNegativeDegreeCentrality(t(NEG1))


[Package xUCINET version 0.0.1.0022 Index]