xNormalize {xUCINET}	R Documentation

Normalizes the values in a network

Description

This function allows different forms of normalizing a network.

Usage


xNormalize(NET1, Type)

Arguments

NET1
A network to be normalized.

Type
The type of symmetrization. The five options are: "Max" (divides the value of each cell value by the maximum of the tie values in a row), "Av" (divides the value of each cell value by the average of the tie values in a row), "Sum" (divides the value of each cell by the sum of the tie values in a row), "Center" (subtracts the mean for a row from each cell value).

Value

The normalized network.

Examples


M1<-matrix(c(1,2,3,4,5,6,7,8,9),3,3)
xNormalize(M1, Type="Max")

[Package xUCINET version 0.0.1.0022 Index]