xPNCentrality {xUCINET}	R Documentation

PN centrality for a negative and positive relation.

Description

An actor's PN centrality is one plus the weighted sum of the PN centrality of the actors they are positively connected to minus a weighted sum of the PN centrality scores of the actors they are negatively connected to. The default weights are used to normalize the scores. The weights are expressed as a beta value (this is because if there are no negative ties the measure is the same as the Hubbell measure) and a weight w for the negative ties No missing data allowed.

Last updated: 1 February 2022.

Usage


xPNCentrality(NET1P, NET1N)

Arguments

NET1P
A one-mode positive network stored as a 'matrix' object.

NET1N
A one-mode negative network stored as a 'matrix' object.

Value

A matrix with PN centrality measures.

References

Everett, M.G. Borgatti, S.P. (2014) Networks containing negative ties. Social Networks 38, 111-120.

Examples


## Undirected one-mode network ## TABLE 4
NEG1<-((Sampson_Monastery$Disesteem+Sampson_Monastery$Dislike+Sampson_Monastery$Blame
        +Sampson_Monastery$NegativeInfluence)>0)*1
POS1<-((Sampson_Monastery$Esteem+Sampson_Monastery$Praise+Sampson_Monastery$PositiveInfluence
        +Sampson_Monastery$LikeT3)>0)*1
POS2<-(POS1+t(POS1))>0
NEG2<-(NEG1+t(NEG1))>0
xPNCentrality(POS2,NEG2)

## Directed one-mode network ## not the same as TABLE 7
xPNCentrality(POS1,NEG1)
xPNCentrality(t(POS1),t(NEG1))


[Package xUCINET version 0.0.1.0022 Index]