xPermuteQAP {xUCINET}	R Documentation

Takes a one-mode network and randomly permutes both rows and columns

Description

This function takes a single one-mode network and randomly reorders (permutes) both rows and columns. This function is used for QAP-based analysis.

Last updated: 23 April 2022.

Usage


xPermuteQAP(NET0)

Arguments

NET0
A network (currently only adjacency matrices are supported).

Value

A permuted adjacency matrix, where rows and columns are reodered in the same way.

Author(s)

Filip Agneessens filipagneessens2@gmail.com

Examples


MAT1<-matrix(c(2,4,2,3,3,1,3,4,1),3,3)
xPermuteQAP(MAT1)

[Package xUCINET version 0.0.1.0022 Index]