xQAPCorrelation {xUCINET}	R Documentation

Performs a correlation test using a classic (and a permutation-based) significance test between two network variables

Description

This function takes two networks among the same nodes and returns the correlation coefficient("CorrCoef") and the classic (two-tailed) significance test ("ClassicP_2T"). If NPERM (number of permutations) is set to a value higher than 0 (default is 1000), a quadratic assignment permutations-based (QAP) significance is also provided, which will provide information about the correlations coefficients for the permuted networks (including the histogram). A higher value than 1000 is generally advisable. Both the one-tailed ("Prop_SmallerThan" or "Prop_GreaterThan") and the two-tailed ("Prop_MoreExtreme") test are given. Note that cases with missing values are included in the permutation and this can impact the degrees of freedom for the correlations obtained for the permutations. For example, suppose the same 20% of cases are missing for both variables (i.e., unit non-response), then permuting these missings will mean that a lot more cases will have a missing for at least one of both variables.

Last updated: 23 April 2022.

Usage


xQAPCorrelation(NET1, NET2, NPERM = 1000, Directed = TRUE, Loops = FALSE)

Arguments

NET1
A first network.

NET2
A second network, among the same nodes as the first network.

NPERM
A positive integer value indicating the number of permutations to be used for the statistical test. If NPERM is set to NULL, NA or 0, no permutation-based significance test is performed.

Directed
Whether the analysis should consider ties in both directions (node i to j and j to i) as separate cases. Default is TRUE.

Loops
Whether to allow self-loops.

Value

Prints a matrix with at least the following values: "NumberOfCases", "NumberOfValid1", "NumberOfValid2", "NumberOfValidBoth", "CorrCoef", "ClassicP_2T", and in case a permutation test is performed, in addition: Mean, StandardDev, Min_Perm, Max_Perm, Prop_SmallerThan, Prop_GreaterThan, Prop_MoreExtreme, Prop_SmallerThanABS, Prop_GreaterThanABS, NumberOfPermut. When using "<-" creates a list object with these respective values as separate elements.

Author(s)

Filip Agneessens filipagneessens2@gmail.com

Examples


X1<-matrix(c(2,4,2,3,3,1,3,4,1),3,3)
X2<-matrix(c(4,1,5,2,1,4,2,2,3),3,3)
RESULS1<-xQAPCorrelation(X1,X2, NPERM=5000)
RESULS1
xQAPCorrelation(X1,X2,NPERM=0)

[Package xUCINET version 0.0.1.0022 Index]