xReachBetweennessCentrality {xUCINET}	R Documentation

Calculates reach betweenness centrality.

Description

Calculates betweenness centrality for a one-mode or two-mode network, i.e., the number (proportion) of times a node 'i' is in between others for paths of length k or less (defined by 'kReach'). kReach has to be a single value. For directed networks is consider the "in-out" connections around node i, i.e., j->i->k, which are of length k or less. Note: Undirected counts are counted in both directions j->i->k and k->i->j. No missing data are allowed.

Usage


xReachBetweennessCentrality(NET1, kReach = 2)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

kReach
A single value indicating the maximum distances being considered.

Value

A matrix with reach betweenness centrality values for a specific k (kReach).

Examples


## Undirected one-mode network
xReachBetweennessCentrality(Padgett_FlorentineFamilies$Marriage)

## Directed one-mode network
xReachBetweennessCentrality(Krackhardt_HighTech$Advice)


[Package xUCINET version 0.0.1.0022 Index]