xReachCentrality {xUCINET}	R Documentation

Calculate kReach centrality.

Description

Calculates the kReach centrality for a one-mode or two-mode network, i.e., the number/proportion of nodes reachable in k steps (if Cumulative=TRUE), or in exactly k steps if Cumulative=FALSE. k is defined with the argument kReach, which can be a single value or a set of values. The Loops for the geodesic distance matrix is by default set to NA. Missing data should be indicated by 'NA'.

Last updated: 1 February 2022.

Usage


xReachCentrality(
  NET1,
  kReach = 2,
  Loops = FALSE,
  Normalize = FALSE,
  Cumulative = TRUE
)

Arguments

NET1
A one-mode or two-mode network stored as a 'matrix' object.

kReach
A single value or vector contain kReach values.

Loops
Logical statement, whether Loops should be included (default=FALSE).

Normalize
Whether to return the normalized value or raw counts.

Cumulative
Whether to consider nodes reached in exactly k steps (FALSE) or in <=k steps (TRUE).

Value

A matrix with kReach centrality values.

Examples


## Undirected one-mode network
xReachCentrality(Padgett_FlorentineFamilies$Marriage, kReach=c(1:6))

## Directed one-mode network (out-reach followed by in-reach)
xReachCentrality(Krackhardt_HighTech$Advice, kReach=c(1:6))
xReachCentrality(t(Krackhardt_HighTech$Advice), kReach=c(1:6))


[Package xUCINET version 0.0.1.0022 Index]