xReciprocity {xUCINET}	R Documentation

Calculates the level of reciprocity in a binary, one-mode network.

Description

This function takes a one-mode network and provides information related to the level of reciprocity in the network. The basis for calculating the level of reciprocity are the number of mutual (M), asymmetric (A) and null (N) dyads. This is better known as the dyad census or MAN distribution. We can either count the number of dyads connected by a tie (which may or may not be reciprocated) and calculate the proportion of dyads that have reciprocated ties (the dyad based method) or we can count the number of arcs (directed edges) and calculate the proportion of arcs that are reciprocated (the arc based method). Note that the Loops entries (self-nominations) are always ignored for this function and that two-mode data are not allowed. The output generates:

Number and proportion of valid dyads. Dyads where either cell i,j or j,i are missing (indicated by values 'NA' in the matrix) are ignored.

The density for the network, i.e. the proportion of ties with values 1 (considering only those dyads that have a valid value in both directions),

The number of mutual dyads (where (i->j)=1 and (j->i)=1),

The number of asymmetric dyads (where (i->j)=1 and (j->i)=0, or (i->j)=0 and (j->i)=1),

The number of null dyads (where (i->j)=0 and (j->i)=0),

The dyad based reciprocity: M/(M+A)

The arc based reciprocity: 2M/(2M+A), which can be compared to the density

Usage


xReciprocity(NET1)

Arguments

NET1
A one-mode network stored as a 'matrix' object.#'

Value

A vector with a series of different measures. See Description.

References

Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

See Also

xDensity(), xTransitivity(), xCyclicality()

Examples


## A network with 4 valid dyads, where the density for the values in these 4 dyads is: 5/8.
NET1<-matrix(c(0,1,NA,0,
      1,0,1,NA,
      1,0,1,1,
      0,NA,1,0),4,4,byrow=TRUE)
xReciprocity(NET1)

## Krackhardt's high-tech managers:
xReciprocity(Krackhardt_HighTech$Advice)
xReciprocity(Krackhardt_HighTech$Friendship)
xReciprocity(Krackhardt_HighTech$ReportTo)

## To show that the arc reciprocity index would be similar to the density for a
## random network, let's generate 100 networks of size 50 by 50 with density .3:
## Store the results for each generated network in RES1:
RES1<-matrix(NA,100,8)
## Now generate 100 networks use a loop:
for (k1 in 1:100)
  {
  MA1<-matrix(runif(2500)>=.7,50,50)
  RES1[k1,]<-t(xReciprocity(MA1))
  }
## We can check that the density for the generated networks is around .3:
mean(RES1[,3])
## Now, we can check that the arc reciprocity index for the generated graphs.
## This is indeed around .3:
mean(RES1[,8])
## Proving that the arc reciprocity index will be close to the density for a random graph.

[Package xUCINET version 0.0.1.0022 Index]