xStructuralHoles {xUCINET}	R Documentation

Measures to capture structural holes.

Description

This function takes a one-mode network (NET1) and calculates measures of structural equivalence. It returns for each node in the rows of the matrix information about effective size and/or constraint. Both measures also work for valued networks.

Usage


xStructuralHoles(
  NET1,
  Measures = c("EffectiveSize", "Constraint"),
  Loops = FALSE,
  Include = "WholeNetwork",
  Direction = "Out"
)

Arguments

NET1
A one-mode network dataset stored as an object of class "matrix".

Measures
The choice of output to be provided. This can be "EffectiveSize", and/or "Constraint".

Loops
Whether to include the Loops (default is to ignore the Loops).

Include
Whether to consider the whole network ("WholeNetwork") or only the egonetwork ("EgoNetwork") around the node.

Direction
How to define the alters for a node (at the moment only outgoing choices).

Value

A matrix containing different measures (columns) for each node (row).

Examples


## Consider the Friendship network for Krackhardt's high-tech managers.
xStructuralHoles(Padgett_FlorentineFamilies$Marriage)
## Consider the Friendship network for Krackhardt's high-tech managers.
xStructuralHoles(Krackhardt_HighTech$Friendship)

[Package xUCINET version 0.0.1.0022 Index]