xTransitivity {xUCINET}	R Documentation

Calculates the level of transitivity in a binary, one-mode network.

Description

This function takes a one-mode network and provides information related to the level of transitivity in the network. For an undirected network this is also sometimes referred to as clustering. The basis for calculating the level of transitivity is the number of transitivity triplets (where i->j=1, j->k=1 and i->k=1) divided by the number of opportunities for a transitivity triplet (where i->j=1 and j->k=1, but where i->k could be either 0 or 1). Note that the Loops entries (self-nominations) are always ignored for this function and that two-mode data are not allowed. Triplets where information on any of the three cells (i,j), (j,k) or (i,k) are missing (indicated by values 'NA' in the matrix) are ignored. The output generates:

Number and proportion of valid triplets, i.e., where (i,j), (j,k) and (i,k) are either 0 or 1, but not NA.

The density for the network, i.e. the proportion of ties with values 1 (considering only those triplets that are valid),

The number of triplets belonging to any of the unique 8 different types of triplets,

The Transitivity Index: (ij=1 & jk=1 & ik=1)/((ij=1 & jk=1 & ik=1)+(ij=1 & jk=1 & ik=0)), which can be compared to the density.

Usage


xTransitivity(NET1)

Arguments

NET1
A one-mode network stored as a 'matrix' object.

Value

A vector with a series of measures. See Description.

References

Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

See Also

xDensity(), xReciprocity(), xCyclicality()

Examples


## A network with 6 valid triplets out of 4*3*2 = 0.25, where the density is 0.722
## and the index is 1/(1+2)=.33.
NET1<-matrix(c(0,1,NA,0,
      1,0,1,NA,
      1,0,1,1,
      0,NA,1,0),4,4,byrow=TRUE)
xTransitivity(NET1)

## Transitivity for Krackhardt's high-tech managers (index of 0.66, 0.46 and 0.00 respectively):
xTransitivity(Krackhardt_HighTech$Advice)
xTransitivity(Krackhardt_HighTech$Friendship)
xTransitivity(Krackhardt_HighTech$ReportTo)

## To show that the transitivity index would be similar to the density for a
## random network, let's generate 100 networks of size 50 by 50 with density .3:
## Store the results for each generated network in RES1:
RES1<-matrix(NA,100,12)
## Now generate 100 networks use a loop:
for (k1 in 1:100)
  {
  MA1<-matrix(runif(2500)>=.7,50,50)
  RES1[k1,]<-t(xTransitivity(MA1))
  }
## We can check that the density for the generated networks is around .3:
mean(RES1[,3])
## Now, we can check that the transitivity index for the generated graphs.
## This is indeed around .3:
mean(RES1[,12])
## Proving that the transitivity index will be close to the density for a random graph.

[Package xUCINET version 0.0.1.0022 Index]