xTwoModeToBipartite {xUCINET}	R Documentation

Transforms a two-mode network into a bipartite network

Description

This function takes a two-mode network stored as a matrix and transforms it into a bipartite network where the row and columns consists of both actors and events.

Usage


xTwoModeToBipartite(NET2, First = "RowNodes")

Arguments

NET2
A single two-mode network stored in an R object as a matrix.

First
Whether to first put RowNodes or ColumnNodes.

Value

A one-mode bipartite network

Examples


xTwoModeToBipartite(Davis_SouthernWomen$Attendance)

[Package xUCINET version 0.0.1.0022 Index]