xTwoModeToOneMode {xUCINET}	R Documentation

Transforms a two-mode network into a one-mode network (one-mode projection).

Description

This function takes a two-mode network stored as a matrix and transforms it into a one-mode network (i.e., a one-mode projection). The type of one-mode projections for nodes of mode 1 based on two-mode data involving ties between mode 1 and mode 2 nodes are:

"CrossProd": the number of common connections of nodes of the other mode. In the case of a valued network, this give the sum of the product of the tie strength to other nodes of the other modes. For z[i,j]=(Sum[k](x[i,k]*x[j,k]))

"CrossProdDMin": the number of common connections of nodes of the other mode divided by the maximum possible given the degree of both nodes. In the case of a valued network, this give the sum of the product of the tie strength to other nodes of the other modes. For z[i,j]=(Sum[k](x[i,k]*x[j,k]))/(Sum[k](min(x[i,k],x[j,k])))

"ExactMatch": the relation between two nodes of a mode is calculated as the number of times the two nodes make the same choices and non-choices. In the case of a valued network, the same strength of tie to the other mode. For z[i,j]=(Sum[k](x[i,k]=x[j,k])). This can also be seen as the absolute difference between 2.

"Jaccard": the number of common connections to nodes of the other mode, divided by the total number of connections to nodes of mode B by either (or both). For z[i,j]=(Sum[k](x[i,k]*x[j,k]))/(Sum[k](x[i,k]=1 AND/OR xjk=1))

"CrossMin": the strength of a tie between two nodes of a mode is the sum of the minimum of both ties. For z[i,j]=(Sum[k](min(x[i,k],x[j,k])))

"Correlation" Pearson's product-moment correlation.

"Covariance" Mean-centered cross products: Sxy/n - SxSy/n^2

Usage


xTwoModeToOneMode(NET2, Measure = "CrossProd")

Arguments

NET2
A single two-mode network stored in an R object as a matrix.

Measure
Character string, indicating which type of transformation to use.

Value

A one-mode projection

Examples


xTwoModeToOneMode(Davis_SouthernWomen$Attendance)

[Package xUCINET version 0.0.1.0022 Index]