xValuedTieComposition {xUCINET}	R Documentation

Provides different measures about the outgoing ties for nodes on a valued network

Description

Takes an object of class 'matrix', which is a valued network, and calculates different measures:

the sum of the number of valid alters, i.e., the number of alters for which information is available (not equal to NA),

the sum of tie strengths to alters (Measures="SumStrength"),

the average tie strengths to alters relative to the number of valid cases (Measures="AvStrength"),

the standard deviation of tie strengths to alters (Measures="SDStrength"),

the median of tie strengths to alters (Measures="MedianStrength"),

the range of tie strengths to alters, i.e., the maximum minus the minimum value (Measures="RangeStrength"),

the minimum tie strengths to alters (Measures="MinStrength"),

the maximum tie strengths to alters (Measures="MaxStrength"),

the interquartile for tie strengths to alters (Measures="IQRStrength"),

the first quartile for tie strengths to alters (Measures="Q1Strength"),

the third quartile for tie strengths to alters (Measures="Q3Strength"),

the IQV for tie strengths to alters (Measures="IQVStrength"),

the distribution of tie strengths to alters (Measures="DistrStrength").

Usage


xValuedTieComposition(
  NET1,
  Measures = c("SumStrength", "AvStrength", "SDStrength", "MedianStrength",
    "RangeStrength", "MinStrength", "MaxStrength", "IQRStrength", "Q1Strength",
    "Q3Strength", "IQVStrength", "DistrStrength"),
  Loops = FALSE,
  Mode = "OneMode",
  Exclude0 = FALSE,
  NrCategoriesIQV = 1
)

Arguments

NET1
A matrix stored as an object of class matrix

Measures
The measures to be calculated about the set of networks. Options are: "SumStrength","AvStrength","SDStrength","MedianStrength","RangeStrength","MinStrength","MaxStrength","IQRStrength","Q1Strength","Q3Strength","IQVStrength", and "DistrStrength".

Loops
Logical, whether the diagonal values should be ignored (default setting is FALSE, i.e., to ignore the diagonal values)

Mode
Whether the networks is one-mode "OneMode" or two-mode networks "TwoMode".

Exclude0
Whether to exclude values 0 from the analysis (default is FALSE).

NrCategoriesIQV
An option value to be used for IQV strength, in case there are some additional values/categories that should be considered, but might not be present in the data. A value 1 means that the observed number of categories should be used.

Value

A matrix with different measures (columns) for each node (row)

References

Chapter 8 in Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.

Examples


## Applied to the valued network for Zachary's Karate club:
xValuedTieComposition(Zachary_KarateClub$Strength, Exclude0=TRUE, NrCategoriesIQV = 20)
## See Chapter 8 in Borgatti, Everett, Johnson and Agneessens (2022)

[Package xUCINET version 0.0.1.0022 Index]