xWalkTrap {xUCINET}	R Documentation

Provides the membership of nodes for the community detection using the walktrap algorithm.

Description

Using an undirected one-mode network (NET1), this function calculates the solution(s) for community detection membership using the walktrap algorithm and graphically shows the result. The output is a matrix providing membership for communities with (k) groups. This number of communities can be identified beforehand in the optional argument (NOC). This can be a single value or a series of valued (as a vector). Values should be at least as large as the number of components in a network. If no values are defined in (NOC), then the solution with the largest value is given (again, where the number of communities is at least as large as the number of components). This function relies on igraph's cluster_walktrap() function.

Usage


xWalkTrap(
  NET1,
  NOC = NULL,
  vertex.shape = NULL,
  vertex.label.cex = NULL,
  layout = NULL
)

Arguments

NET1
An undirected one-mode network stored as a 'matrix' object.

NOC
A number or vector containing the number of communities (NOC) to be defined. This needs to be at least as large as the number of components in the network.

vertex.shape
An option argument to specify the shapes of the nodes. Needs to be a vector of length equal to the number of nodes.

vertex.label.cex
The size of the font used for the labels (default when NULL is cex=1).

layout
The layout option used. See igraph. (default when NULL is layout_nicely, which tries to find the optimal layout).

Value

A vector or matrix with membership for each node (row).

References

Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE. igraph package - https://igraph.org

See Also

xGirvanNewman(), xFastGreedy(), xLouvainMethod(), xLabelPropagation(), igraph::cluster_walktrap()

Examples


## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
xWalkTrap(Zachary_KarateClub$Connection, vertex.label.cex=.7)
## Add attribute file related to the club split:
## Define the shapes to be used
UseShapes<-c("circle","rectangle")
## Define the vector for nodes -
ZKAttrClub<-(Zachary_KarateClub$Attributes$Club>1)+1
xWalkTrap(Zachary_KarateClub$Connection, vertex.shape=UseShapes[ZKAttrClub],
          vertex.label.cex=.7)
xWalkTrap(Zachary_KarateClub$Connection, NOC=c(3,4,5),vertex.shape=UseShapes[ZKAttrClub],
          vertex.label.cex=.7)

## A second dataset:
xWalkTrap(Kapferer_TailorShop$SociationalT1, vertex.label.cex=.7, layout="layout_with_fr")

[Package xUCINET version 0.0.1.0022 Index]